Struct stacktrace::Trace

pub struct Trace<E> {
    pub err: E,
    pub stacktrace: StackInfo,
}

An error 'annotated' with a stack trace.

Fields

err

stacktrace

Methods

impl<E> Trace<E>

fn new(e: E) -> Self

Construct a new Trace struct, pairing the provided value with a new stack trace.

fn result<A, F: From<E>>(res: Result<A, E>) -> Result<A, Trace<F>>

Convert a Result<A, E> to a Result<A, Trace<E>> by inserting a new stack trace. This also applys any relevant instances of From.

Use this if you want to trace the error now, but need to handle the trace and error later.

Example

use stacktrace::Trace;

struct MyError(String);

fn error() -> Result<usize, MyError> {
  Err(MyError("oh no!".to_owned()))
}

fn deal_with_it_later() -> Result<(), Trace<MyError>> {
  // the returned stack trace will lead to this line in this function.
  let num = try!(Trace::result(error()));
  println!("Got result {}!", num);
  Ok(())
}

fn trace<A, F: From<E>>(res: Result<A, Trace<E>>) -> Result<A, Trace<F>>

Apply any instances of From to the wrapped error, while keeping the previous stack trace.

Use this in combination with the try! macro to convert errors before returning them.

Example

use stacktrace::Trace;

struct ErrorA(String);

struct ErrorB(ErrorA);
impl From<ErrorA> for ErrorB {
  fn from(err: ErrorA) -> ErrorB { ErrorB(err) }
}

fn error() -> Result<usize, Trace<ErrorA>> {
  Err(Trace::new(ErrorA("oh no!".to_owned())))
}

fn deal_with_it_later() -> Result<(), Trace<ErrorB>> {
  let num = try!(Trace::trace(error()));
  println!("Got {}!", num);
  Ok(())
}

Note

For the curious reader, the reason we can't derive a generic instance of From<Trace<E>> for Trace<F> is because it would conflict with the generic impl From<T> for T in std.